dashboard page improvements
